requirements.txt
streamlit==1.41.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.36
firebase-admin==6.6.0
pyrebase4==4.8.0
beautifulsoup4==4.12.3
requests==2.32.3
thefuzz==0.22.1
python-Levenshtein==0.26.1
Pillow==11.1.0
plotly==5.24.1
pandas==2.2.3
python-dotenv==1.0.1
